Develop

+++ src/Service/Application/Internals/Calculator.cs
@@ -3,10 +3,10 @@
 internal static class Calculator
 {
     // Function to calculate the net insurance cost for a specific coverage
-    internal static double CalculateInsuranceCost(double rate, int capital) =>
+    internal static decimal CalculateInsuranceCost(decimal rate, decimal capital) =>
         rate * capital;
 
     // Function to calculate the total net insurance cost for multiple coverages
-    internal static double CalculateTotalInsuranceCost(IEnumerable<(double Rate, int Capitals)> data) =>
+    internal static decimal CalculateTotalInsuranceCost(IEnumerable<(decimal Rate, decimal Capitals)> data) =>
         data.Sum(dataItem => CalculateInsuranceCost(dataItem.Rate, dataItem.Capitals));
 }
